+++ finesse/src/App.js
@@ -1,8 +1,8 @@
 
-import ProductList from "./pages/ProductList";
+import Cart from "./pages/Cart";
 
 const App = () => {
-    return <ProductList/>;
+    return <Cart/>;
   };
   
   export default App;
